Add JoinGeometry node and remove all old nodes
