--- UnityGeometryGraph/Assets/GeometryGraph/Editor/Graph/Nodes/DisplayValueNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8d10a69a37ef4b5dae6a769bc0d61eb9
-timeCreated: 1630579008
