--- UnityGeometryGraph/Assets/GeometryGraph/Editor/Graph/Nodes/Float.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 67ee0b003af34bc8bad2531978b017b7
-timeCreated: 1630579538
