--- UnityGeometryGraph/Assets/GeometryGraph/Editor/Graph/Nodes/Float/FloatInputNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2945e9d00e8f4e9b81ae5de9b9b9cf54
-timeCreated: 1630571912
