--- UnityGeometryGraph/Assets/GeometryGraph/Editor/Graph/Nodes/Float/FloatMathNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9e5e759c8a6d4c6a94bbad51d24ee486
-timeCreated: 1630574635
