--- UnityGeometryGraph/Assets/GeometryGraph/Runtime/Graph/Node/DisplayValueNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 0599a25428d44283a9233083b6085478
-timeCreated: 1631792536
