--- UnityGeometryGraph/Assets/GeometryGraph/Runtime/Graph/Node/FloatInputNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4def21a0db6d493d8f58c6af6d1b3041
-timeCreated: 1631791791
