--- UnityGeometryGraph/Assets/GeometryGraph/Runtime/Graph/Node/FloatMathNode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 24d912385ea44e84a6641a7fe6890c65
-timeCreated: 1631791949
