+++ UnityGeometryGraph/Assets/New Geometry Graph.geometrygraph
@@ -1,1 +1,1 @@
-{"AssetGuid":"17521bed217a0ec4b85261f723566f2c","IsBlackboardVisible":true,"GraphVersion":"1.0.0","RuntimeGraphData":{"id":0},"nodes":[{"GUID":"7690675b-973f-442b-a198-3407d42c226e","Type":"GeometryGraph.Editor.OutputNode","DrawState":{"Position":{"serializedVersion":"2","x":495.0,"y":245.0,"width":135.0,"height":75.0},"Expanded":true},"NodeData":"{}","PortData":["ad5bccda-bfb2-4083-b74a-c01f6dd14ccd"]},{"GUID":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","Type":"GeometryGraph.Editor.GeometryObjectPropertyNode","DrawState":{"Position":{"serializedVersion":"2","x":268.0,"y":254.0,"width":200.0,"height":150.0},"Expanded":true},"NodeData":"{\"propertyGuid\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\"}","PortData":["9bc489fd-222d-45d5-82a8-a8bc45aaad9a"]}],"edges":[{"Input":"7690675b-973f-442b-a198-3407d42c226e","Output":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","InputPort":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd","OutputPort":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","InputCapacity":0,"OutputCapacity":1}],"serializedProperties":[{"Type":"GeometryGraph.Editor.GeometryObjectProperty","Data":"{\"$type\":\"GeometryGraph.Editor.GeometryObjectProperty, GeometryGraph.Editor\",\"GUID\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\",\"Type\":\"GeometryObject\",\"DisplayName\":\"Geometry Object\",\"ReferenceName\":\"GeometryObject_95BBD0F4\",\"OverrideReferenceName\":null,\"Hidden\":false,\"ShortGuid\":\"95BBD0F4\"}"},{"Type":"GeometryGraph.Editor.GeometryObjectProperty","Data":"{\"$type\":\"GeometryGraph.Editor.GeometryObjectProperty, GeometryGraph.Editor\",\"GUID\":\"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc\",\"Type\":\"GeometryObject\",\"DisplayName\":\"Geometry Object 2\",\"ReferenceName\":\"GeometryObject_F66C000E\",\"OverrideReferenceName\":null,\"Hidden\":false,\"ShortGuid\":\"F66C000E\"}"}],"references":{"version":1,"00000000":{"type":{"class":"RuntimeGraphObjectData","ns":"GeometryGraph.Runtime","asm":"GeometryGraph.Runtime"},"data":{"Guid":"beeff82b-d259-479e-b884-51798df80ee4","Connections":[{"OutputGuid":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","InputGuid":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd"}],"Properties":[{"Guid":"0c1fcba2-987c-4067-a462-a4fa32931d1f","ReferenceName":"GeometryObject_95BBD0F4","DisplayName":"Geometry Object","Type":0},{"Guid":"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc","ReferenceName":"GeometryObject_F66C000E","DisplayName":"Geometry Object 2","Type":0}],"serializedRuntimeNodes":[{"Guid":"7690675b-973f-442b-a198-3407d42c226e","Type":"GeometryGraph.Runtime.Graph.OutputNode","CustomData":"","Ports":[{"Guid":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd","Type":2,"Direction":0}]},{"Guid":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","Type":"GeometryGraph.Runtime.Graph.GeometryObjectPropertyNode","CustomData":"{\"p\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\"}","Ports":[{"Guid":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","Type":2,"Direction":1}]}]}},"00000001":{"type":{"class":"Terminus","ns":"UnityEngine.DMAT","asm":"FAKE_ASM"},"data":{}}}}
+{"AssetGuid":"17521bed217a0ec4b85261f723566f2c","IsBlackboardVisible":true,"GraphVersion":"1.0.0","RuntimeGraphData":{"id":0},"nodes":[{"GUID":"7690675b-973f-442b-a198-3407d42c226e","Type":"GeometryGraph.Editor.OutputNode","DrawState":{"Position":{"serializedVersion":"2","x":626.0,"y":252.0,"width":135.0,"height":75.0},"Expanded":true},"NodeData":"{}","PortData":["ad5bccda-bfb2-4083-b74a-c01f6dd14ccd"]},{"GUID":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","Type":"GeometryGraph.Editor.GeometryObjectPropertyNode","DrawState":{"Position":{"serializedVersion":"2","x":215.0,"y":226.0,"width":155.0,"height":75.0},"Expanded":true},"NodeData":"{\"propertyGuid\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\"}","PortData":["9bc489fd-222d-45d5-82a8-a8bc45aaad9a"]},{"GUID":"852b0741-8e9b-4313-ac5b-7daa2d3e2a66","Type":"GeometryGraph.Editor.GeometryObjectPropertyNode","DrawState":{"Position":{"serializedVersion":"2","x":205.0,"y":320.0,"width":165.0,"height":75.0},"Expanded":true},"NodeData":"{\"propertyGuid\":\"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc\"}","PortData":["31f9f17a-68d5-44d0-967f-378765c664e7"]},{"GUID":"addef34b-d5bf-4c15-8aeb-424e3158e812","Type":"GeometryGraph.Editor.JoinGeometryNode","DrawState":{"Position":{"serializedVersion":"2","x":458.0,"y":252.0,"width":140.0,"height":99.0},"Expanded":true},"NodeData":"{}","PortData":["4e84f955-5592-41d3-880b-733e1a6357c2","bce7c660-42ba-489b-bbba-9ff275a46892","eae37f8c-b893-4baf-88e4-d3dbaaf88326"]}],"edges":[{"Input":"7690675b-973f-442b-a198-3407d42c226e","Output":"addef34b-d5bf-4c15-8aeb-424e3158e812","InputPort":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd","OutputPort":"eae37f8c-b893-4baf-88e4-d3dbaaf88326","InputCapacity":0,"OutputCapacity":1},{"Input":"addef34b-d5bf-4c15-8aeb-424e3158e812","Output":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","InputPort":"4e84f955-5592-41d3-880b-733e1a6357c2","OutputPort":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","InputCapacity":0,"OutputCapacity":1},{"Input":"addef34b-d5bf-4c15-8aeb-424e3158e812","Output":"852b0741-8e9b-4313-ac5b-7daa2d3e2a66","InputPort":"bce7c660-42ba-489b-bbba-9ff275a46892","OutputPort":"31f9f17a-68d5-44d0-967f-378765c664e7","InputCapacity":0,"OutputCapacity":1}],"serializedProperties":[{"Type":"GeometryGraph.Editor.GeometryObjectProperty","Data":"{\"$type\":\"GeometryGraph.Editor.GeometryObjectProperty, GeometryGraph.Editor\",\"GUID\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\",\"Type\":\"GeometryObject\",\"DisplayName\":\"Geometry Object\",\"ReferenceName\":\"GeometryObject_95BBD0F4\",\"OverrideReferenceName\":null,\"Hidden\":false,\"ShortGuid\":\"95BBD0F4\"}"},{"Type":"GeometryGraph.Editor.GeometryObjectProperty","Data":"{\"$type\":\"GeometryGraph.Editor.GeometryObjectProperty, GeometryGraph.Editor\",\"GUID\":\"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc\",\"Type\":\"GeometryObject\",\"DisplayName\":\"Geometry Object 2\",\"ReferenceName\":\"GeometryObject_F66C000E\",\"OverrideReferenceName\":null,\"Hidden\":false,\"ShortGuid\":\"F66C000E\"}"}],"references":{"version":1,"00000000":{"type":{"class":"RuntimeGraphObjectData","ns":"GeometryGraph.Runtime","asm":"GeometryGraph.Runtime"},"data":{"Guid":"beeff82b-d259-479e-b884-51798df80ee4","Connections":[{"OutputGuid":"eae37f8c-b893-4baf-88e4-d3dbaaf88326","InputGuid":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd"},{"OutputGuid":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","InputGuid":"4e84f955-5592-41d3-880b-733e1a6357c2"},{"OutputGuid":"31f9f17a-68d5-44d0-967f-378765c664e7","InputGuid":"bce7c660-42ba-489b-bbba-9ff275a46892"}],"Properties":[{"Guid":"0c1fcba2-987c-4067-a462-a4fa32931d1f","ReferenceName":"GeometryObject_95BBD0F4","DisplayName":"Geometry Object","Type":0},{"Guid":"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc","ReferenceName":"GeometryObject_F66C000E","DisplayName":"Geometry Object 2","Type":0}],"serializedRuntimeNodes":[{"Guid":"7690675b-973f-442b-a198-3407d42c226e","Type":"GeometryGraph.Runtime.Graph.OutputNode","CustomData":"","Ports":[{"Guid":"ad5bccda-bfb2-4083-b74a-c01f6dd14ccd","Type":2,"Direction":0}]},{"Guid":"7f48c197-ea37-423c-a17c-e2d6e406d4f9","Type":"GeometryGraph.Runtime.Graph.GeometryObjectPropertyNode","CustomData":"{\"p\":\"0c1fcba2-987c-4067-a462-a4fa32931d1f\"}","Ports":[{"Guid":"9bc489fd-222d-45d5-82a8-a8bc45aaad9a","Type":2,"Direction":1}]},{"Guid":"852b0741-8e9b-4313-ac5b-7daa2d3e2a66","Type":"GeometryGraph.Runtime.Graph.GeometryObjectPropertyNode","CustomData":"{\"p\":\"0d84fa80-e5a6-4f59-8cbf-101a4630cdfc\"}","Ports":[{"Guid":"31f9f17a-68d5-44d0-967f-378765c664e7","Type":2,"Direction":1}]},{"Guid":"addef34b-d5bf-4c15-8aeb-424e3158e812","Type":"GeometryGraph.Runtime.Graph.JoinGeometryNode","CustomData":"","Ports":[{"Guid":"4e84f955-5592-41d3-880b-733e1a6357c2","Type":2,"Direction":0},{"Guid":"bce7c660-42ba-489b-bbba-9ff275a46892","Type":2,"Direction":0},{"Guid":"eae37f8c-b893-4baf-88e4-d3dbaaf88326","Type":2,"Direction":1}]}]}},"00000001":{"type":{"class":"Terminus","ns":"UnityEngine.DMAT","asm":"FAKE_ASM"},"data":{}}}}
